requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
trimesh>=4.0.0
shapely>=2.0.0
mapbox-earcut>=1.0.0
google-cloud-aiplatform>=1.38.0
google-generativeai>=0.8.0
